feat: Distinguish internal and external activities
Introduces a `type` field to the `Activity` interface to differentiate between internal and external activities. This allows for more accurate cost tracking and reporting by enabling the capture of `companyCost` for external activities.

The `Booking` interface has been updated to remove the `external_activity` type, as this is now handled by the new `type` field on `Activity`. The logic for calculating `itemCost` has also been adjusted to accommodate this distinction.